Shrink gmo_message
On Linux x86_64, sizeof(gmo_message) == 32:

    8 bytes: const char* message_.begin_
    8 bytes: const char* message_.end_
    4 bytes: std::uint32_t hash_
    4 bytes: padding

The size of gmo_message matters a bit because the all_diagnostic_infos
table contains many gmo_message objects.

Replace message_ with a pointer and an int, shrinking gmo_message to 16
bytes:

    8 bytes: const char* message_
    4 bytes: int message_length_
    4 bytes: std::uint32_t hash_

+++ test/test-gmo.cpp
@@ -112,8 +112,8 @@ TEST(test_gmo, missing_translation_gives_original_string) {
   gmo_file p(big_endian_two_string_gmo_file);
 
   const char *message = "does not exist";
-  std::string_view translated =
-      p.find_translation(gmo_message(message, std::strlen(message)));
+  std::string_view translated = p.find_translation(
+      gmo_message(message, narrow_cast<int>(std::strlen(message))));
   EXPECT_EQ(translated.data(), message);
 }
