+++ tools/dump-gmo.cpp
@@ -5,6 +5,7 @@
 #include <iostream>
 #include <quick-lint-js/file.h>
 #include <quick-lint-js/gmo.h>
+#include <quick-lint-js/narrow-cast.h>
 
 int main(int argc, char** argv) {
   using namespace quick_lint_js;
@@ -23,8 +24,9 @@ int main(int argc, char** argv) {
     std::string_view translated = gmo.translated_string_at(i);
     std::cerr << original << "\n  -> " << translated << '\n';
 
-    std::string_view translated_by_lookup = gmo.find_translation(
-        quick_lint_js::gmo_message(original.data(), original.size()));
+    std::string_view translated_by_lookup =
+        gmo.find_translation(quick_lint_js::gmo_message(
+            original.data(), narrow_cast<int>(original.size())));
     if (translated_by_lookup != translated) {
       std::cerr << "    !!! error: lookup returned instead: "
                 << translated_by_lookup << '\n';
